backend/requirements.txt
pandas
scikit-learn
xgboost
numpy
joblib
fastapi
uvicorn
pydantic 
sqlalchemy
pymysql
python-dotenv 
matplotlib
seaborn